+++ R/reexport-rlang.R
@@ -3,3 +3,9 @@
 #' @export
 #' @name set_names
 rlang::set_names
+
+#' @title Execute a function
+#' @keywords NULL
+#' @export
+#' @name exec
+rlang::exec